requirements.txt
fastapi
uvicorn
openai
pinecone
python-dotenv
jinja2
python-multipart
PyPDF2
python-docx
lxml
pandas
tqdm
httpx